change to int user id

+++ vk-bot-main/src/main/java/com/barinthecityshow/bot/handler/BotRequestHandler.java
@@ -1,5 +1,5 @@
 package com.barinthecityshow.bot.handler;
 
 public interface BotRequestHandler {
-    void handle(Long userId, String msg);
+    void handle(Integer userId, String msg);
 }
